fix no_default_deps
This improves performance of nasm in release sanitizer build.

e.g.
obj/third_party/dav1d/dav1d_asm/itx16_avx2.o is assembled in 6.3s with
this CL, but it took 19.8s without this CL in release asan build on
linux. It is improved from 150.3s to 8.9s on Windows.

Bug: 1374606
Change-Id: I2ace0a0c14f213e5bd5cd0dd9ff9f05eec0e9945