nasm: don't depend on canonicalize_file_name(3)
It is a glibc-specific function that has a portable replacement, and
nasm already has code to fall back to this portable replacement.
Having nasm depend on the glibc-specific function is one of the
obstacles to chromium building with other libcs, so getting rid of
it is nice.

Note that the relevant config header (config/config-linux.h) is an
indirect result of running autoconf + configure on a developer
workstation when rolling a new nasm version, so those config
headers do not necessarily reflect the environment chromium is being
built in when we ship source tarballs. To work around that, this
change adds a post-processing step that corrects configure's output,
and adds a note to run it as part of the nasm uprev process.

Since I was touching nasm anyway, this change also ports
generate_nasm_sources.py from python2 to python3.

Bug: 1380656
Change-Id: I59dc3bda4a119cd440d05674550aa5c3af6b2390